Update OceanPatcher
wow, I missed a lot the first time around.

basically, whoever made the shipwrecks was insane and I had to match that insanity to fix it (stairs and fences were not connected/ shaped correctly AT ALL)

this was really rough to do. updates to the layers comes in the next commit